pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "hiero-did-sdk-python"
version = "0.1.0"
description = "A Python SDK for creating and resolving Decentralized Identifiers (DIDs) on Hiero/Hedera networks using the did:hedera method."
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
authors = [
    { name = "ClearBrick Contributors" },
]
keywords = ["hedera", "hiero", "did", "decentralized-identity", "w3c", "verifiable-credentials", "blockchain"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Security :: Cryptography",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

dependencies = [
    "hiero-sdk-python>=0.2.0",
    "py-multibase>=2.0.0",
    "requests>=2.31.0,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<10",
    "pytest-cov>=5.0.0,<8",
    "ruff>=0.4.0,<1",
    "mypy>=1.10.0,<2",
    "types-requests>=2.31.0,<3",
    "pre-commit>=3.0.0,<5",
]

[project.urls]
Homepage = "https://github.com/baseramp/hiero-did-sdk-python"
Repository = "https://github.com/baseramp/hiero-did-sdk-python"
Issues = "https://github.com/baseramp/hiero-did-sdk-python/issues"
Documentation = "https://github.com/baseramp/hiero-did-sdk-python#readme"

[tool.hatch.build.targets.wheel]
packages = ["src/hiero_did"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "D",
    "E",
    "F",
    "I",
    "B",
    "UP",
    "SIM",
    "PERF",
    "ARG",
    "RET",
    "RSE",
]
ignore = ["D212", "E501", "B008", "D100", "D104"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D102", "D103", "ARG002"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring a live Hiero network (deselect with '-m \"not integration\"')",
]

[tool.coverage.run]
source = ["hiero_did"]

[tool.coverage.report]
show_missing = true
fail_under = 70
